CGame::GetSyncCounter() should return size_t

+++ src/game.h
@@ -325,7 +325,7 @@ public:
   inline bool                                            GetIsLobbyOrMirror() const { return !m_GameLoading && !m_GameLoaded; }
   inline bool                                            GetIsLobbyStrict() const { return !m_IsMirror && !m_GameLoading && !m_GameLoaded; }
   inline bool                                            GetIsRestored() const { return m_RestoredGame != nullptr; }
-  inline uint32_t                                        GetSyncCounter() const { return m_SyncCounter; }
+  inline size_t                                          GetSyncCounter() const { return m_SyncCounter; }
   uint8_t                                                GetMaxEqualizerDelayFrames() const { return m_MaxPingEqualizerDelayFrames; }
   uint8_t                                                CalcMaxEqualizerDelayFrames() const;
   int64_t                                                GetActiveLatency() const;
